requirements.txt
requests
python-telegram-bot==20.3
os
